requirements.txt
fastapi==0.89.1
uvicorn==0.20.0
requests==2.28.2
python-multipart==0.0.5
python-decouple==3.7